v0.5

+++ Customer.java
@@ -1,10 +1,8 @@
 package com.company;
 
-
 public class Customer {
-
 	//datafield tiplerini degistirebilirsiniz
-	double arrivalTime; //musteri gelis zamani-
+	int arrivalTime; //musteri gelis zamani-
 	//bekleme zamanini hesaplamada kullanabilirsiniz
-	double removalTime;
+	int removalTime;
 }
